+++ Source/NSManagedObjectContext+Extensions.swift
@@ -24,22 +24,40 @@ extension NSManagedObjectContext {
     /// Describes a child managed object context.
     public typealias ChildContext = NSManagedObjectContext
 
-    /**
-     Attempts to commit unsaved changes to registered objects in the context.
-     This function is performed in a block on the context's queue. If the context has no changes,
-     then this function returns immediately and the completion block is not called.
+    /// Describes the result type for saving a managed object context.
+    public typealias SaveResult = Result<NSManagedObjectContext, Error>
 
-     - parameter wait:       If `true` (the default), saves synchronously. If `false`, saves asynchronously.
-     - parameter completion: The closure to be executed when the save operation completes.
-     */
-    public func save(wait: Bool = true, completion: ((SaveResult) -> Void)? = nil) {
+    /// Attempts to **asynchronously** commit unsaved changes to registered objects in the context.
+    /// This function is performed in a block on the context's queue. If the context has no changes,
+    /// then this function returns immediately and the completion block is not called.
+    ///
+    /// - Parameter completion: The closure to be executed when the save operation completes.
+    public func saveAsync(completion: ((SaveResult) -> Void)? = nil) {
+        _save(wait: false, completion: completion)
+    }
+
+    /// Attempts to **synchronously** commit unsaved changes to registered objects in the context.
+    /// This function is performed in a block on the context's queue. If the context has no changes,
+    /// then this function returns immediately and the completion block is not called.
+    ///
+    /// - Parameter completion: The closure to be executed when the save operation completes.
+    public func saveSync(completion: ((SaveResult) -> Void)? = nil) {
+        _save(wait: true, completion: completion)
+    }
+
+    /// Attempts to commit unsaved changes to registered objects in the context.
+    ///
+    /// - Parameter wait: If `true`, saves synchronously. If `false`, saves asynchronously.
+    /// - Parameter completion: The closure to be executed when the save operation completes.
+    private func _save(wait: Bool, completion: ((SaveResult) -> Void)? = nil) {
+
         let block = {
             guard self.hasChanges else { return }
             do {
                 try self.save()
-                completion?(.success)
+                completion?(.success(self))
             } catch {
-                completion?(.failure(error as NSError))
+                completion?(.failure(error))
             }
         }
         wait ? performAndWait(block) : perform(block)